modularize labels instead of tags into inventory
