+++ ereuse_devicehub/templates/ereuse_devicehub/base_site.html
@@ -181,7 +181,7 @@
     <li class="nav-heading">Utils</li>
 
     <li class="nav-item">
-      <a class="nav-link collapsed" href="{{ url_for('inventory.taglist')}}">
+      <a class="nav-link collapsed" href="{{ url_for('label.label_list')}}">
         <i class="bi bi-tags"></i>
         <span>Tags</span>
       </a>
