+++ ereuse_devicehub/templates/label/label_list.html
@@ -20,7 +20,7 @@
           <!-- Bordered Tabs -->
 
             <div class="btn-group dropdown ml-1">
-              <a href="{{ url_for('inventory.tag_add')}}" type="button" class="btn btn-primary">
+              <a href="{{ url_for('label.tag_add')}}" type="button" class="btn btn-primary">
                 <i class="bi bi-plus"></i>
                 Create Named Tag
                 <span class="caret"></span>
@@ -28,7 +28,7 @@
             </div>
 
             <div class="btn-group dropdown ml-1" uib-dropdown="">
-              <a href="{{ url_for('inventory.tag_unnamed_add')}}" type="button" class="btn btn-primary">
+              <a href="{{ url_for('label.tag_unnamed_add')}}" type="button" class="btn btn-primary">
                 <i class="bi bi-plus"></i>
                 Create UnNamed Tag
                 <span class="caret"></span>
@@ -52,7 +52,7 @@
                   <tbody>
                     {% for tag in tags %}
                     <tr>
-                      <td><a href="{{ url_for('inventory.tag_details', id=tag.id) }}">{{ tag.id }}</a></td>
+                      <td><a href="{{ url_for('label.label_details', id=tag.id) }}">{{ tag.id }}</a></td>
                       <td>{% if tag.provider %}Unnamed tag {% else %}Named tag{% endif %}</td>
                       <td>{{ tag.get_provider }}</td>
                       <td>
